+++ Runtime/Utils.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Diagnostics;
+using UnityEngine;
+
+
+namespace CinderUtils {
+    public static class Utils {
+        #region CONSTANTS
+        public const string MY_DEBUG = "MY_DEBUG";
+
+        #endregion
+
+
+
+
+
+
+    }
+}